+++ tests/macros/tdumpastgen.nim
@@ -1,5 +1,6 @@
 discard """
-nimout: '''nnkStmtList.newTree(
+nimout: '''
+nnkStmtList.newTree(
   nnkVarSection.newTree(
     nnkIdentDefs.newTree(
       newIdentNode("x"),
@@ -26,20 +27,73 @@ nimout: '''nnkStmtList.newTree(
       newCommentStmtNode("This is a docstring"),
       nnkCommand.newTree(
         newIdentNode("echo"),
-        newLit("bar")
+        newLit("Hello, World!")
+      ),
+      nnkCommand.newTree(
+        newIdentNode("echo"),
+        newLit("something \"quoted\"")
+      )
+    )
+  ),
+  nnkCall.newTree(
+    newIdentNode("callNilLit"),
+    newNilLit()
+  ),
+  nnkAsgn.newTree(
+    nnkDotExpr.newTree(
+      newIdentNode("x"),
+      newIdentNode("y")
+    ),
+    nnkObjConstr.newTree(
+      newIdentNode("MyType"),
+      nnkExprColonExpr.newTree(
+        newIdentNode("u1"),
+        nnkUInt64Lit.newTree(
+        )
+      ),
+      nnkExprColonExpr.newTree(
+        newIdentNode("u2"),
+        nnkUInt32Lit.newTree(
+        )
       )
     )
   )
-)'''
+)
+'''
 """
 
-# disabled; can't work as the output is done by the compiler
-
 import macros
+import stdtest/unittest_light
 
 dumpAstGen:
   var x = baz.create(56)
-
   proc foo() =
     ## This is a docstring
-    echo "bar"
+    echo "Hello, World!"
+    echo "something \"quoted\""
+
+  callNilLit(nil)
+  x.y = MyType(u1: 123'u64, u2: 321'u32)
+
+macro myQuoteAst(arg: untyped): untyped = astGen(arg)
+
+static:
+  let myAst = myQuoteAst:
+    var x = baz.create(56)
+    proc foo() =
+      ## This is a docstring
+      echo "Hello, World!"
+      echo "something \"quoted\""
+
+    callNilLit(nil)
+    x.y = MyType(u1: 123'u64, u2: 321'u32)
+  assertEquals myAst.repr, """
+
+var x = baz.create(56)
+proc foo() =
+  ## This is a docstring
+  echo "Hello, World!"
+  echo "something \"quoted\""
+
+callNilLit(nil)
+x.y = MyType(u1: 123'u64, u2: 321'u32)"""