+++ changelog.md
@@ -210,6 +210,8 @@
   at which progress callbacks are called.
 
 - Added `os.copyFileToDir`
+- Added `macros.newLit(NimNode)` overload.
+- Added `macros.astGen(NimNode)`.
 
 ## Language changes